Create util file
Move util functions to util file to be shared by all models
